Fix Show instances for Opcode causing infinite recursion
The following instance causes infinite recursion for `Opcode' ()`, since
`concrete = void`, i.e. `show opcode` becomes `show (concrete opcode)`
becomes `show (show (concrete opcode))`, etc. ad infinitum.

```haskell
-instance {-# OVERLAPPABLE #-} Show a => Show (Opcode' a) where
-  show opcode = show (concrete opcode) <> showParam opcode
```

The overlapping instance for `Opcode' ()` was supposed to have been the
base case for this recursion, but it apparently never triggered. Because
the Show instance was not tested, and the test suite only uses the Show
instance when there are errors, an error was not caused for a long time,
since the only updates have been version bumps since the regression.

- Remove FlexibleInstances
- Remove OVERLAPPING instances for Show
- Remove orphan instance for Opcode' ()
- Test `show` for `Opcode' ()`

This fixes #3.